requirements.txt
click==8.1.3
Flask==2.2.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pkg-resources==0.0.0
python-dotenv==0.21.1
typing-extensions==4.5.0
Werkzeug==2.2.3
zipp==3.15.0
